Lines now generate points and points can parameterize!!
Now to draw them. This is a major advancement in the program. I now know
I can draw straight lines and apply any complex function to them and get
the homotopy.

Next steps include writing the point grid to be a collection of lines
(and circles probably will need to be included at some point). Then to
write a way to serialize that data so it can be read by the matplotlib
animation library.